Substitute spaces for underscore in class value (ugly hack)
MAI degree has subjects taught in different colleges, so some subjects have
spaces in the class value, like so: "Aula 200, URV".

This commit substitutes those spaces with underscores, so that when classes
are iterated in domini.py@_afegeix_classes the class is correctly parsed
into its five components.

Being a hack, this method needs to be revised and cleaned up in the near
future.